init with typescript
