+++ server/src/config/database.ts
@@ -12,23 +12,5 @@ const sequelize = new Sequelize(database, username, password, {
   },
 });
 
-let users = [
-  {
-    id: 1,
-    name: `sanuja`,
-    address: `kalaeliya`,
-  },
-  {
-    id: 2,
-    name: `amal`,
-    address: `wattala`,
-  },
-  {
-    id: 3,
-    name: `kamal`,
-    address: `nigambo`,
-  },
-];
 
-exports.users = users;
 exports.sequelize = sequelize;
